Improve phrasing of Plum exercise instructions
The existing instructions are ambiguous about whether you should change the text so it matches the initial preview (i.e. all on one line), or change it so the four lines that are present in the code are also present in the preview. Hopefully this phrasing is clearer.

I just helped someone thru the tutorial, and they ran straight into this ambiguity, so it is a real source of confusion.